+++ soil_analysis/domain/valueobject/geocode/yahoo.py
